Remove old dimension type form component
